==> 1 - compute_features.R <==
#### Compute features for 10-,20-,30-minutely and 1-hourly data ####
# we save these .rds files separately because they are very large

library(dplyr)
library(tsibble)
library("readxl")
library(fpp3)
library(zoo)
library(stats)
library(foreach)
library(doParallel)
library(lightgbm)
library(distributional)


setwd(dirname(rstudioapi::getActiveDocumentContext()$path)) # set this to your working directory

min10 <- readRDS(file = "min10.rds")
min20 <- readRDS(file = "min20.rds")
min30 <- readRDS(file = "min30.rds")
hr1 <- readRDS(file = "hr1.rds")

# compute features on data set, 10-minutely
min10 <- min10 %>% mutate(
  `WMA2` = slider::slide_dbl(Wind_Speed, mean,
                             .before = 3, .after = -1, .complete = TRUE),
  `WMA3` = slider::slide_dbl(Wind_Speed, mean,
                             .before = 4, .after = -1, .complete = TRUE),
  `WMA4` = slider::slide_dbl(Wind_Speed, mean,
                             .before = 5, .after = -1, .complete = TRUE),
  `WMA5` = slider::slide_dbl(Wind_Speed, mean,
                             .before = 6, .after = -1, .complete = TRUE),
  `WMA6` = slider::slide_dbl(Wind_Speed, mean,
                             .before = 7, .after = -1, .complete = TRUE),
  `WMSD2` = slider::slide_dbl(Wind_Speed, sd,
                              .before = 3, .after = -1, .complete = TRUE),
  `WMSD3` = slider::slide_dbl(Wind_Speed, sd,
                              .before = 4, .after = -1, .complete = TRUE),
  `WMSD4` = slider::slide_dbl(Wind_Speed, sd,
                              .before = 5, .after = -1, .complete = TRUE),
  `WMSD5` = slider::slide_dbl(Wind_Speed, sd,
                              .before = 6, .after = -1, .complete = TRUE),
  `WMSD6` = slider::slide_dbl(Wind_Speed, sd,
                              .before = 7, .after = -1, .complete = TRUE),
  `PMA2` = slider::slide_dbl(Energy, mean,
                             .before = 3, .after = -1, .complete = TRUE),
  `PMA3` = slider::slide_dbl(Energy, mean,
                             .before = 4, .after = -1, .complete = TRUE),
  `PMA4` = slider::slide_dbl(Energy, mean,
                             .before = 5, .after = -1, .complete = TRUE),
  `PMA5` = slider::slide_dbl(Energy, mean,
                             .before = 6, .after = -1, .complete = TRUE),
  `PMA6` = slider::slide_dbl(Energy, mean,
                             .before = 7, .after = -1, .complete = TRUE),
  `PMSD2` = slider::slide_dbl(Energy, sd,
                              .before = 3, .after = -1, .complete = TRUE),
  `PMSD3` = slider::slide_dbl(Energy, sd,
                              .before = 4, .after = -1, .complete = TRUE),
  `PMSD4` = slider::slide_dbl(Energy, sd,
                              .before = 5, .after = -1, .complete = TRUE),
  `PMSD5` = slider::slide_dbl(Energy, sd,
                              .before = 6, .after = -1, .complete = TRUE),
  `PMSD6` = slider::slide_dbl(Energy, sd,
                              .before = 7, .after = -1, .complete = TRUE),
  `lag_wind1` = lag(Wind_Speed, 1),
  `lag_wind2` = lag(Wind_Speed, 2),
  `lag_wind3` = lag(Wind_Speed, 3),
  `lag_wind4` = lag(Wind_Speed, 4),
  `lag_wind5` = lag(Wind_Speed, 5),
  `lag_wind6` = lag(Wind_Speed, 6),
  `lag_wind7` = lag(Wind_Speed, 7),
  `lag_wind8` = lag(Wind_Speed, 8),
  `lag_wind9` = lag(Wind_Speed, 9),
  `lag_wind10` = lag(Wind_Speed, 10),
  `lag_wind11` = lag(Wind_Speed, 11),
  `lag_Energy1` = lag(Energy, 1),
  `lag_Energy2` = lag(Energy, 2),
  `lag_Energy3` = lag(Energy, 3),
  `lag_Energy4` = lag(Energy, 4),
  `lag_Energy5` = lag(Energy, 5),
  `lag_Energy6` = lag(Energy, 6),
  `lag_Energy7` = lag(Energy, 7),
  `lag_Energy8` = lag(Energy, 8),
  `lag_Energy9` = lag(Energy, 9),
  `lag_Energy10` = lag(Energy, 10),
  `lag_Energy11` = lag(Energy, 11),
  `is_q1` = as.integer(quarter(Time)==1),
  `is_q2` = as.integer(quarter(Time)==2),
  `is_q3` = as.integer(quarter(Time)==3),
  `is_00` = as.integer(hour(Time) == 0),
  `is_01` = as.integer(hour(Time) == 1),
  `is_02` = as.integer(hour(Time) == 2),
  `is_03` = as.integer(hour(Time) == 3),
  `is_04` = as.integer(hour(Time) == 4),
  `is_05` = as.integer(hour(Time) == 5),
  `is_06` = as.integer(hour(Time) == 6),
  `is_07` = as.integer(hour(Time) == 7),
  `is_08` = as.integer(hour(Time) == 8),
  `is_09` = as.integer(hour(Time) == 9),
  `is_10` = as.integer(hour(Time) == 10),
  `is_11` = as.integer(hour(Time) == 11),
  `is_12` = as.integer(hour(Time) == 12),
  `is_13` = as.integer(hour(Time) == 13),
  `is_14` = as.integer(hour(Time) == 14),
  `is_15` = as.integer(hour(Time) == 15),
  `is_16` = as.integer(hour(Time) == 16),
  `is_17` = as.integer(hour(Time) == 17),
  `is_18` = as.integer(hour(Time) == 18),
  `is_19` = as.integer(hour(Time) == 19),
  `is_20` = as.integer(hour(Time) == 20),
  `is_21` = as.integer(hour(Time) == 21),
  `is_22` = as.integer(hour(Time) == 22)
)

min10 <- min10 %>% mutate(
  `WMA2_l1` = lag(WMA2,1),
  `WMA3_l1` = lag(WMA3,1),
  `WMA4_l1` = lag(WMA4,1),
  `WMA5_l1` = lag(WMA5,1),
  `WMA6_l1` = lag(WMA6,1),
  `WMSD2_l1` = lag(WMSD2,1),
  `WMSD3_l1` = lag(WMSD3,1),
  `WMSD4_l1` = lag(WMSD4,1),
  `WMSD5_l1` = lag(WMSD5,1),
  `WMSD6_l1` = lag(WMSD6,1),
  `PMA2_l1` = lag(PMA2, 1),
  `PMA3_l1` = lag(PMA3, 1),
  `PMA4_l1` = lag(PMA4, 1),
  `PMA5_l1` = lag(PMA5, 1),
  `PMA6_l1` = lag(PMA6, 1),
  `PMSD2_l1` = lag(PMSD2, 1),
  `PMSD3_l1` = lag(PMSD3, 1),
  `PMSD4_l1` = lag(PMSD4, 1),
  `PMSD5_l1` = lag(PMSD5, 1),
  `PMSD6_l1` = lag(PMSD6, 1),
  `WMA2_l2` = lag(WMA2,2),
  `WMA3_l2` = lag(WMA3,2),
  `WMA4_l2` = lag(WMA4,2),
  `WMA5_l2` = lag(WMA5,2),
  `WMA6_l2` = lag(WMA6,2),
  `WMSD2_l2` = lag(WMSD2,2),
  `WMSD3_l2` = lag(WMSD3,2),
  `WMSD4_l2` = lag(WMSD4,2),
  `WMSD5_l2` = lag(WMSD5,2),
  `WMSD6_l2` = lag(WMSD6,2),
  `PMA2_l2` = lag(PMA2, 2),
  `PMA3_l2` = lag(PMA3, 2),
  `PMA4_l2` = lag(PMA4, 2),
  `PMA5_l2` = lag(PMA5, 2),
  `PMA6_l2` = lag(PMA6, 2),
  `PMSD2_l2` = lag(PMSD2, 2),
  `PMSD3_l2` = lag(PMSD3, 2),
  `PMSD4_l2` = lag(PMSD4, 2),
  `PMSD5_l2` = lag(PMSD5, 2),
  `PMSD6_l2` = lag(PMSD6, 2),
  `WMA2_l3` = lag(WMA2,3),
  `WMA3_l3` = lag(WMA3,3),
  `WMA4_l3` = lag(WMA4,3),
  `WMA5_l3` = lag(WMA5,3),
  `WMA6_l3` = lag(WMA6,3),
  `WMSD2_l3` = lag(WMSD2,3),
  `WMSD3_l3` = lag(WMSD3,3),
  `WMSD4_l3` = lag(WMSD4,3),
  `WMSD5_l3` = lag(WMSD5,3),
  `WMSD6_l3` = lag(WMSD6,3),
  `PMA2_l3` = lag(PMA2, 3),
  `PMA3_l3` = lag(PMA3, 3),
  `PMA4_l3` = lag(PMA4, 3),
  `PMA5_l3` = lag(PMA5, 3),
  `PMA6_l3` = lag(PMA6, 3),
  `PMSD2_l3` = lag(PMSD2, 3),
  `PMSD3_l3` = lag(PMSD3, 3),
  `PMSD4_l3` = lag(PMSD4, 3),
  `PMSD5_l3` = lag(PMSD5, 3),
  `PMSD6_l3` = lag(PMSD6, 3),
  `WMA2_l4` = lag(WMA2,4),
  `WMA3_l4` = lag(WMA3,4),
  `WMA4_l4` = lag(WMA4,4),
  `WMA5_l4` = lag(WMA5,4),
  `WMA6_l4` = lag(WMA6,4),
  `WMSD2_l4` = lag(WMSD2,4),
  `WMSD3_l4` = lag(WMSD3,4),
  `WMSD4_l4` = lag(WMSD4,4),
  `WMSD5_l4` = lag(WMSD5,4),
  `WMSD6_l4` = lag(WMSD6,4),
  `PMA2_l4` = lag(PMA2, 4),
  `PMA3_l4` = lag(PMA3, 4),
  `PMA4_l4` = lag(PMA4, 4),
  `PMA5_l4` = lag(PMA5, 4),
  `PMA6_l4` = lag(PMA6, 4),
  `PMSD2_l4` = lag(PMSD2, 4),
  `PMSD3_l4` = lag(PMSD3, 4),
  `PMSD4_l4` = lag(PMSD4, 4),
  `PMSD5_l4` = lag(PMSD5, 4),
  `PMSD6_l4` = lag(PMSD6, 4),
  `WMA2_l5` = lag(WMA2,5),
  `WMA3_l5` = lag(WMA3,5),
  `WMA4_l5` = lag(WMA4,5),
  `WMA5_l5` = lag(WMA5,5),
  `WMA6_l5` = lag(WMA6,5),
  `WMSD2_l5` = lag(WMSD2,5),
  `WMSD3_l5` = lag(WMSD3,5),
  `WMSD4_l5` = lag(WMSD4,5),
  `WMSD5_l5` = lag(WMSD5,5),
  `WMSD6_l5` = lag(WMSD6,5),
  `PMA2_l5` = lag(PMA2, 5),
  `PMA3_l5` = lag(PMA3, 5),
  `PMA4_l5` = lag(PMA4, 5),
  `PMA5_l5` = lag(PMA5, 5),
  `PMA6_l5` = lag(PMA6, 5),
  `PMSD2_l5` = lag(PMSD2, 5),
  `PMSD3_l5` = lag(PMSD3, 5),
  `PMSD4_l5` = lag(PMSD4, 5),
  `PMSD5_l5` = lag(PMSD5, 5),
  `PMSD6_l5` = lag(PMSD6, 5)
)

# compute features on data set, 20-minutely
min20 <- min20 %>% mutate(
  `WMA2` = slider::slide_dbl(Wind_Speed, mean,
                             .before = 3, .after = -1, .complete = TRUE),
  `WMA3` = slider::slide_dbl(Wind_Speed, mean,
                             .before = 4, .after = -1, .complete = TRUE),
  `WMA4` = slider::slide_dbl(Wind_Speed, mean,
                             .before = 5, .after = -1, .complete = TRUE),
  `WMA5` = slider::slide_dbl(Wind_Speed, mean,
                             .before = 6, .after = -1, .complete = TRUE),
  `WMA6` = slider::slide_dbl(Wind_Speed, mean,
                             .before = 7, .after = -1, .complete = TRUE),
  `WMSD2` = slider::slide_dbl(Wind_Speed, sd,
                              .before = 3, .after = -1, .complete = TRUE),
  `WMSD3` = slider::slide_dbl(Wind_Speed, sd,
                              .before = 4, .after = -1, .complete = TRUE),
  `WMSD4` = slider::slide_dbl(Wind_Speed, sd,
                              .before = 5, .after = -1, .complete = TRUE),
  `WMSD5` = slider::slide_dbl(Wind_Speed, sd,
                              .before = 6, .after = -1, .complete = TRUE),
  `WMSD6` = slider::slide_dbl(Wind_Speed, sd,
                              .before = 7, .after = -1, .complete = TRUE),
  `PMA2` = slider::slide_dbl(Energy, mean,
                             .before = 3, .after = -1, .complete = TRUE),
  `PMA3` = slider::slide_dbl(Energy, mean,
                             .before = 4, .after = -1, .complete = TRUE),
  `PMA4` = slider::slide_dbl(Energy, mean,
                             .before = 5, .after = -1, .complete = TRUE),
  `PMA5` = slider::slide_dbl(Energy, mean,
                             .before = 6, .after = -1, .complete = TRUE),
  `PMA6` = slider::slide_dbl(Energy, mean,
                             .before = 7, .after = -1, .complete = TRUE),
  `PMSD2` = slider::slide_dbl(Energy, sd,
                              .before = 3, .after = -1, .complete = TRUE),
  `PMSD3` = slider::slide_dbl(Energy, sd,
                              .before = 4, .after = -1, .complete = TRUE),
  `PMSD4` = slider::slide_dbl(Energy, sd,
                              .before = 5, .after = -1, .complete = TRUE),
  `PMSD5` = slider::slide_dbl(Energy, sd,
                              .before = 6, .after = -1, .complete = TRUE),
  `PMSD6` = slider::slide_dbl(Energy, sd,
                              .before = 7, .after = -1, .complete = TRUE),
  `lag_wind1` = lag(Wind_Speed, 1),
  `lag_wind2` = lag(Wind_Speed, 2),
  `lag_wind3` = lag(Wind_Speed, 3),
  `lag_wind4` = lag(Wind_Speed, 4),
  `lag_wind5` = lag(Wind_Speed, 5),
  `lag_wind6` = lag(Wind_Speed, 6),
  `lag_wind7` = lag(Wind_Speed, 7),
  `lag_wind8` = lag(Wind_Speed, 8),
  `lag_Energy1` = lag(Energy, 1),
  `lag_Energy2` = lag(Energy, 2),
  `lag_Energy3` = lag(Energy, 3),
  `lag_Energy4` = lag(Energy, 4),
  `lag_Energy5` = lag(Energy, 5),
  `lag_Energy6` = lag(Energy, 6),
  `lag_Energy7` = lag(Energy, 7),
  `lag_Energy8` = lag(Energy, 8),
  `is_q1` = as.integer(quarter(Time)==1),
  `is_q2` = as.integer(quarter(Time)==2),
  `is_q3` = as.integer(quarter(Time)==3),
  `is_00` = as.integer(hour(Time) == 0),
  `is_01` = as.integer(hour(Time) == 1),
  `is_02` = as.integer(hour(Time) == 2),
  `is_03` = as.integer(hour(Time) == 3),
  `is_04` = as.integer(hour(Time) == 4),
  `is_05` = as.integer(hour(Time) == 5),
  `is_06` = as.integer(hour(Time) == 6),
  `is_07` = as.integer(hour(Time) == 7),
  `is_08` = as.integer(hour(Time) == 8),
  `is_09` = as.integer(hour(Time) == 9),
  `is_10` = as.integer(hour(Time) == 10),
  `is_11` = as.integer(hour(Time) == 11),
  `is_12` = as.integer(hour(Time) == 12),
  `is_13` = as.integer(hour(Time) == 13),
  `is_14` = as.integer(hour(Time) == 14),
  `is_15` = as.integer(hour(Time) == 15),
  `is_16` = as.integer(hour(Time) == 16),
  `is_17` = as.integer(hour(Time) == 17),
  `is_18` = as.integer(hour(Time) == 18),
  `is_19` = as.integer(hour(Time) == 19),
  `is_20` = as.integer(hour(Time) == 20),
  `is_21` = as.integer(hour(Time) == 21),
  `is_22` = as.integer(hour(Time) == 22)
)

min20 <- min20 %>% mutate(
  `WMA2_l1` = lag(WMA2,1),
  `WMA3_l1` = lag(WMA3,1),
  `WMA4_l1` = lag(WMA4,1),
  `WMA5_l1` = lag(WMA5,1),
  `WMA6_l1` = lag(WMA6,1),
  `WMSD2_l1` = lag(WMSD2,1),
  `WMSD3_l1` = lag(WMSD3,1),
  `WMSD4_l1` = lag(WMSD4,1),
  `WMSD5_l1` = lag(WMSD5,1),
  `WMSD6_l1` = lag(WMSD6,1),
  `PMA2_l1` = lag(PMA2, 1),
  `PMA3_l1` = lag(PMA3, 1),
  `PMA4_l1` = lag(PMA4, 1),
  `PMA5_l1` = lag(PMA5, 1),
  `PMA6_l1` = lag(PMA6, 1),
  `PMSD2_l1` = lag(PMSD2, 1),
  `PMSD3_l1` = lag(PMSD3, 1),
  `PMSD4_l1` = lag(PMSD4, 1),
  `PMSD5_l1` = lag(PMSD5, 1),
  `PMSD6_l1` = lag(PMSD6, 1),
  `WMA2_l2` = lag(WMA2,2),
  `WMA3_l2` = lag(WMA3,2),
  `WMA4_l2` = lag(WMA4,2),
  `WMA5_l2` = lag(WMA5,2),
  `WMA6_l2` = lag(WMA6,2),
  `WMSD2_l2` = lag(WMSD2,2),
  `WMSD3_l2` = lag(WMSD3,2),
  `WMSD4_l2` = lag(WMSD4,2),
  `WMSD5_l2` = lag(WMSD5,2),
  `WMSD6_l2` = lag(WMSD6,2),
  `PMA2_l2` = lag(PMA2, 2),
  `PMA3_l2` = lag(PMA3, 2),
  `PMA4_l2` = lag(PMA4, 2),
  `PMA5_l2` = lag(PMA5, 2),
  `PMA6_l2` = lag(PMA6, 2),
  `PMSD2_l2` = lag(PMSD2, 2),
  `PMSD3_l2` = lag(PMSD3, 2),
  `PMSD4_l2` = lag(PMSD4, 2),
  `PMSD5_l2` = lag(PMSD5, 2),
  `PMSD6_l2` = lag(PMSD6, 2)
)

# compute features on data set, 30-minutely
min30 <- min30 %>% mutate(
  `WMA2` = slider::slide_dbl(Wind_Speed, mean,
                             .before = 3, .after = -1, .complete = TRUE),
  `WMA3` = slider::slide_dbl(Wind_Speed, mean,
                             .before = 4, .after = -1, .complete = TRUE),
  `WMA4` = slider::slide_dbl(Wind_Speed, mean,
                             .before = 5, .after = -1, .complete = TRUE),
  `WMA5` = slider::slide_dbl(Wind_Speed, mean,
                             .before = 6, .after = -1, .complete = TRUE),
  `WMA6` = slider::slide_dbl(Wind_Speed, mean,
                             .before = 7, .after = -1, .complete = TRUE),
  `WMSD2` = slider::slide_dbl(Wind_Speed, sd,
                              .before = 3, .after = -1, .complete = TRUE),
  `WMSD3` = slider::slide_dbl(Wind_Speed, sd,
                              .before = 4, .after = -1, .complete = TRUE),
  `WMSD4` = slider::slide_dbl(Wind_Speed, sd,
                              .before = 5, .after = -1, .complete = TRUE),
  `WMSD5` = slider::slide_dbl(Wind_Speed, sd,
                              .before = 6, .after = -1, .complete = TRUE),
  `WMSD6` = slider::slide_dbl(Wind_Speed, sd,
                              .before = 7, .after = -1, .complete = TRUE),
  `PMA2` = slider::slide_dbl(Energy, mean,
                             .before = 3, .after = -1, .complete = TRUE),
  `PMA3` = slider::slide_dbl(Energy, mean,
                             .before = 4, .after = -1, .complete = TRUE),
  `PMA4` = slider::slide_dbl(Energy, mean,
                             .before = 5, .after = -1, .complete = TRUE),
  `PMA5` = slider::slide_dbl(Energy, mean,
                             .before = 6, .after = -1, .complete = TRUE),
  `PMA6` = slider::slide_dbl(Energy, mean,
                             .before = 7, .after = -1, .complete = TRUE),
  `PMSD2` = slider::slide_dbl(Energy, sd,
                              .before = 3, .after = -1, .complete = TRUE),
  `PMSD3` = slider::slide_dbl(Energy, sd,
                              .before = 4, .after = -1, .complete = TRUE),
  `PMSD4` = slider::slide_dbl(Energy, sd,
                              .before = 5, .after = -1, .complete = TRUE),
  `PMSD5` = slider::slide_dbl(Energy, sd,
                              .before = 6, .after = -1, .complete = TRUE),
  `PMSD6` = slider::slide_dbl(Energy, sd,
                              .before = 7, .after = -1, .complete = TRUE),
  `lag_wind1` = lag(Wind_Speed, 1),
  `lag_wind2` = lag(Wind_Speed, 2),
  `lag_wind3` = lag(Wind_Speed, 3),
  `lag_wind4` = lag(Wind_Speed, 4),
  `lag_wind5` = lag(Wind_Speed, 5),
  `lag_wind6` = lag(Wind_Speed, 6),
  `lag_wind7` = lag(Wind_Speed, 7),
  `lag_Energy1` = lag(Energy, 1),
  `lag_Energy2` = lag(Energy, 2),
  `lag_Energy3` = lag(Energy, 3),
  `lag_Energy4` = lag(Energy, 4),
  `lag_Energy5` = lag(Energy, 5),
  `lag_Energy6` = lag(Energy, 6),
  `lag_Energy7` = lag(Energy, 7),
  `is_q1` = as.integer(quarter(Time)==1),
  `is_q2` = as.integer(quarter(Time)==2),
  `is_q3` = as.integer(quarter(Time)==3),
  `is_00` = as.integer(hour(Time) == 0),
  `is_01` = as.integer(hour(Time) == 1),
  `is_02` = as.integer(hour(Time) == 2),
  `is_03` = as.integer(hour(Time) == 3),
  `is_04` = as.integer(hour(Time) == 4),
  `is_05` = as.integer(hour(Time) == 5),
  `is_06` = as.integer(hour(Time) == 6),
  `is_07` = as.integer(hour(Time) == 7),
  `is_08` = as.integer(hour(Time) == 8),
  `is_09` = as.integer(hour(Time) == 9),
  `is_10` = as.integer(hour(Time) == 10),
  `is_11` = as.integer(hour(Time) == 11),
  `is_12` = as.integer(hour(Time) == 12),
  `is_13` = as.integer(hour(Time) == 13),
  `is_14` = as.integer(hour(Time) == 14),
  `is_15` = as.integer(hour(Time) == 15),
  `is_16` = as.integer(hour(Time) == 16),
  `is_17` = as.integer(hour(Time) == 17),
  `is_18` = as.integer(hour(Time) == 18),
  `is_19` = as.integer(hour(Time) == 19),
  `is_20` = as.integer(hour(Time) == 20),
  `is_21` = as.integer(hour(Time) == 21),
  `is_22` = as.integer(hour(Time) == 22)
)

min30 <- min30 %>% mutate(
  `WMA2_l1` = lag(WMA2,1),
  `WMA3_l1` = lag(WMA3,1),
  `WMA4_l1` = lag(WMA4,1),
  `WMA5_l1` = lag(WMA5,1),
  `WMA6_l1` = lag(WMA6,1),
  `WMSD2_l1` = lag(WMSD2,1),
  `WMSD3_l1` = lag(WMSD3,1),
  `WMSD4_l1` = lag(WMSD4,1),
  `WMSD5_l1` = lag(WMSD5,1),
  `WMSD6_l1` = lag(WMSD6,1),
  `PMA2_l1` = lag(PMA2, 1),
  `PMA3_l1` = lag(PMA3, 1),
  `PMA4_l1` = lag(PMA4, 1),
  `PMA5_l1` = lag(PMA5, 1),
  `PMA6_l1` = lag(PMA6, 1),
  `PMSD2_l1` = lag(PMSD2, 1),
  `PMSD3_l1` = lag(PMSD3, 1),
  `PMSD4_l1` = lag(PMSD4, 1),
  `PMSD5_l1` = lag(PMSD5, 1),
  `PMSD6_l1` = lag(PMSD6, 1)
)

# compute features on data set, 1-hourly
hr1 <- hr1 %>% mutate(
  `WMA2` = slider::slide_dbl(Wind_Speed, mean,
                             .before = 3, .after = -1, .complete = TRUE),
  `WMA3` = slider::slide_dbl(Wind_Speed, mean,
                             .before = 4, .after = -1, .complete = TRUE),
  `WMA4` = slider::slide_dbl(Wind_Speed, mean,
                             .before = 5, .after = -1, .complete = TRUE),
  `WMA5` = slider::slide_dbl(Wind_Speed, mean,
                             .before = 6, .after = -1, .complete = TRUE),
  `WMA6` = slider::slide_dbl(Wind_Speed, mean,
                             .before = 7, .after = -1, .complete = TRUE),
  `WMSD2` = slider::slide_dbl(Wind_Speed, sd,
                              .before = 3, .after = -1, .complete = TRUE),
  `WMSD3` = slider::slide_dbl(Wind_Speed, sd,
                              .before = 4, .after = -1, .complete = TRUE),
  `WMSD4` = slider::slide_dbl(Wind_Speed, sd,
                              .before = 5, .after = -1, .complete = TRUE),
  `WMSD5` = slider::slide_dbl(Wind_Speed, sd,
                              .before = 6, .after = -1, .complete = TRUE),
  `WMSD6` = slider::slide_dbl(Wind_Speed, sd,
                              .before = 7, .after = -1, .complete = TRUE),
  `PMA2` = slider::slide_dbl(Energy, mean,
                             .before = 3, .after = -1, .complete = TRUE),
  `PMA3` = slider::slide_dbl(Energy, mean,
                             .before = 4, .after = -1, .complete = TRUE),
  `PMA4` = slider::slide_dbl(Energy, mean,
                             .before = 5, .after = -1, .complete = TRUE),
  `PMA5` = slider::slide_dbl(Energy, mean,
                             .before = 6, .after = -1, .complete = TRUE),
  `PMA6` = slider::slide_dbl(Energy, mean,
                             .before = 7, .after = -1, .complete = TRUE),
  `PMSD2` = slider::slide_dbl(Energy, sd,
                              .before = 3, .after = -1, .complete = TRUE),
  `PMSD3` = slider::slide_dbl(Energy, sd,
                              .before = 4, .after = -1, .complete = TRUE),
  `PMSD4` = slider::slide_dbl(Energy, sd,
                              .before = 5, .after = -1, .complete = TRUE),
  `PMSD5` = slider::slide_dbl(Energy, sd,
                              .before = 6, .after = -1, .complete = TRUE),
  `PMSD6` = slider::slide_dbl(Energy, sd,
                              .before = 7, .after = -1, .complete = TRUE),
  `lag_wind1` = lag(Wind_Speed, 1),
  `lag_wind2` = lag(Wind_Speed, 2),
  `lag_wind3` = lag(Wind_Speed, 3),
  `lag_wind4` = lag(Wind_Speed, 4),
  `lag_wind5` = lag(Wind_Speed, 5),
  `lag_wind6` = lag(Wind_Speed, 6),
  `lag_wind7` = lag(Wind_Speed, 7),
  `lag_Energy1` = lag(Energy, 1),
  `lag_Energy2` = lag(Energy, 2),
  `lag_Energy3` = lag(Energy, 3),
  `lag_Energy4` = lag(Energy, 4),
  `lag_Energy5` = lag(Energy, 5),
  `lag_Energy6` = lag(Energy, 6),
  `lag_Energy7` = lag(Energy, 7),
  `is_q1` = as.integer(quarter(Time)==1),
  `is_q2` = as.integer(quarter(Time)==2),
  `is_q3` = as.integer(quarter(Time)==3),
  `is_00` = as.integer(hour(Time) == 0),
  `is_01` = as.integer(hour(Time) == 1),
  `is_02` = as.integer(hour(Time) == 2),
  `is_03` = as.integer(hour(Time) == 3),
  `is_04` = as.integer(hour(Time) == 4),
  `is_05` = as.integer(hour(Time) == 5),
  `is_06` = as.integer(hour(Time) == 6),
  `is_07` = as.integer(hour(Time) == 7),
  `is_08` = as.integer(hour(Time) == 8),
  `is_09` = as.integer(hour(Time) == 9),
  `is_10` = as.integer(hour(Time) == 10),
  `is_11` = as.integer(hour(Time) == 11),
  `is_12` = as.integer(hour(Time) == 12),
  `is_13` = as.integer(hour(Time) == 13),
  `is_14` = as.integer(hour(Time) == 14),
  `is_15` = as.integer(hour(Time) == 15),
  `is_16` = as.integer(hour(Time) == 16),
  `is_17` = as.integer(hour(Time) == 17),
  `is_18` = as.integer(hour(Time) == 18),
  `is_19` = as.integer(hour(Time) == 19),
  `is_20` = as.integer(hour(Time) == 20),
  `is_21` = as.integer(hour(Time) == 21),
  `is_22` = as.integer(hour(Time) == 22)
)

saveRDS(min10, file = "min10_wFeatures.rds")
saveRDS(min20, file = "min20_wFeatures.rds")
saveRDS(min30, file = "min30_wFeatures.rds")
saveRDS(hr1, file = "hr1_wFeatures.rds")
